+++ xdef/changelog.md
@@ -3,6 +3,8 @@
 # Version 42.1.8, release-date 2024-06-19
 * Fixed bug in `onIllegalAttr` and `onIllegalText` section of script - code was
   not executed or was not enabled.
+* Corrected report of incorrect error message:
+  `XDEF422 Duplicated script section`
 
 # Version 42.1.7, release-date 2024-06-13
 * Fixed bug when ignoring `whitespace` type restrictions in XML Schema when